Updated remaining file tests to new model
